# doc/api.yaml
openapi:  3.0.3
info: 
  title: Wasa Photo - OpenAPI
  description: |
    API for sharing photos.
    
    Each user will be presented with a stream of photos (images) in reverse chronological order, with information about when each photo was uploaded (date and time) and how many likes and comments it has. The stream is composed by photos from “following” (other users that the user follows). Users can place (and later remove) a “like” to photos from other users. Also, users can add comments to any image (even those uploaded by themself). Only authors can remove their comments. Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about Alice. Alice can decide to remove the ban at any moment. Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse chronological order), how many photos have been uploaded, and the user’s followers and following. Users can change their usernames, upload photos, remove photos, and follow/unfollow other users. Removal of an image will also remove likes and comments.
      A user can search other user profiles via username
  version: 1.0.1
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  parameters:
    username:
      schema:
        description: username schema.
        type: string
        pattern: '^.*?$'
        example: SiweiHu
        minLength: 5
        maxLength: 16
      name: username
      in: path
      required: true
      description: identifier for users
    followUsername:
      schema:
        description: username schema.
        type: string
        pattern: '^.*?$'
        example: SiweiHu
        minLength: 5
        maxLength: 16
      name: followUsername
      in: path
      required: true
      description: identifier for users to follow
    banUsername:
      schema:
        description: username schema.
        type: string
        pattern: '^.*?$'
        example: SiweiHu
        minLength: 5
        maxLength: 16
      name: banUsername
      in: path
      required: true
      description: identifier for users to ban
    likeUsername:
      schema:
        description: username schema.
        type: string
        pattern: '^.*?$'
        example: SiweiHu
        minLength: 5
        maxLength: 16
      name: likeUsername
      in: path
      required: true
      description: identifier for users of like
...
    pid:
      schema: 
        description: Int64 between 0-9999
        type: integer
        format: int64
        example : 999
        readOnly: true
      name: pid
      in: path
      description: A photo unique identifier
      required: true
  schemas:
    User:
      title: User
      description: this objects represents a single user
      type: object
      properties:
        uid:
          description: identifier of the user
          type: integer
          format: int64
          example: 999
        username:
          type: string
          example: SiweiHu
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          readOnly: false
...
    Like:
      title: Like
      description: this object represents a single comment 
      type: object
      properties:
        id:
          description: the identifier of the like
          type: integer
          format: int64
          example: 999
        username:
          description: the username of the user that liked the photo
          type: string
          pattern: '^.*?$'
          example: Leo
          minLength: 3
          maxLength: 16
    Profile:
      title: Profile
      description: represents the user profile
      type: object
      properties:
        username:
          type: string
          pattern: '^.*?$'
          example: SiweiHu
          minLength: 5
          maxLength: 16
        n_photos:
          description: number of photos published by the user
          type: integer
          example: 1
        n_followers:
          description: number of followers of the user
          type: integer
          example: 1
        n_following:
          description: number of other users followed by the user
          type: integer
          example: 1
    Follow:
      title: Follow 
      description: Follow schema.
      type: object
      properties:
        followedUser:
          description: Unique user identifier.
          type: integer
          example: 1
        userId:
          description: Unique user identifier
          type: integer
          example: 1
    Ban:
      title: Ban
      description: Ban schema.
      type: object
      properties:
        bannedUser:
          description: Unique user identifier.
          type: integer
          example: 1
        userId:
          description: Unique user identifier.
          type: integer
          example: 1
...
paths:
...
  /users/{username}/profile:
    parameters:
       - $ref: "#/components/parameters/username"
    
    get:
      security:
        - bearerAuth : []
      tags:
        - user
      operationId: getUserProfile
      summary: returns the profile of the user
      description: | 
        the returned profile shows photos, number of photos uploaded, followers and following
      
      responses:
        '200':
          description: profile informations successfully obtained
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile"}
        "400": 
          description: The request was not compliant with the documentation (eg. missing fields, etc).
        "404": 
          description: User not found 
        "500": 
          description: The server encountered an internal error. Further info in server logs.
...
  /users/{username}/photos/{pid}/likes:
    parameters: 
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/pid"
    
    get:
      security:
        - bearerAuth: []
      tags:
        - photo
      operationId: getPhotoLikes
      summary: get photo's like list
      description: |
          getting all likes of this photo
      responses:
        '200':
          description: likes of the photo retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Like"}
                minItems: 0
                maxItems: 99999
  /users/{username}/photos/{pid}/likes/{likeUsername}:
    parameters: 
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/pid"
      - $ref: "#/components/parameters/likeUsername"
                
    put:
      security:
          - bearerAuth: []
      tags:
        - photo
      operationId: likePhoto
      summary: put a like to a photo
      description: id of the user of like will be added to the collection likes
      responses:
        '201':
          description:  operation successfull
          content:
            application/json:
              schema: { $ref : "#/components/schemas/User"}
              
              
    delete:
      security:
          - bearerAuth: []
      tags:
        - photo
      operationId: unlikePhoto
      summary: remove a like to a photo
      description: id of the user will be deleted from the collection likes
      responses:
        '204':
          description:  operation successfull
  /users/{username}/following/:
    parameters: 
      - $ref: "#/components/parameters/username"
      
    get:
      security:
        - bearerAuth: []
      tags:
        - user
      operationId: listFollowed
      summary: get user's follow list
      description: |
          getting all followers of this user, user objects will be returned
      responses:
        '200':
          description: follow list of the user retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User"}
                minItems: 0
                maxItems: 99999
  /users/{username}/following/{followUsername}: 
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/followUsername"
    put:
      security:
          - bearerAuth: []
      tags:
        - user
      operationId: followUser
      summary: follow a user
      description: id of the user to follow will be added to the collection following
      responses:
        '201':
          description:  operation successfull
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Follow"}
              
    delete:
      security:
          - bearerAuth: []
      tags:
        - user
      operationId: unfollowUser
      summary: remove a user from following
      description: id of the user will be deleted from the collection following
      responses:
        '204':
          description:  operation successfull
...
  /users/{username}/banned/:
    parameters: 
      - $ref: "#/components/parameters/username"
      
    get:
      security:
        - bearerAuth: []
      tags:
        - user
      operationId: listBanned
      summary: get user's ban list
      description: |
          getting all banned user of this user
      responses:
        '200':
          description: ban list of the user retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User"}
                minItems: 0
                maxItems: 99999
  /users/{username}/banned/{banUsername}:
    parameters: 
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/banUsername"
    
    put:
      security:
          - bearerAuth: []
      tags:
        - user
      operationId: banUser
      summary: ban a user
      description: id of the user to ban will be added to the collection banned
      responses:
        '201':
          description:  operation successfull
          content:
            application/json:
              schema: { $ref : "#/components/schemas/Ban"}
              
    delete:
      security:
          - bearerAuth: []
      tags:
        - user
      operationId: unbanUser
      summary: remove a user from banned
      description: id of the user will be deleted from the collection banned
      responses:
        '204':
          description:  operation successfull

              
              
              
              
              
              
              
                
              
              
# 